core: input.tsx
