+++ static/app/components/core/input/index.tsx
@@ -1,1 +1,1 @@
-export {Input, type InputProps} from './input';
+export {Input, inputStyles, type InputProps, type InputStylesProps} from './input';
