+++ static/app/components/core/interactionStateLayer/interactionStateLayer.stories.tsx
